+++ tests/conftest.py
@@ -1,4 +1,5 @@
 import gc
+import os
 
 import pytest
 
@@ -7,3 +8,9 @@ import pytest
 def force_gc():
     yield
     gc.collect()
+
+
+@pytest.fixture(scope="session")
+def findoc_model_id() -> str:
+    """Identifier of the Financial Document model, supplied through an env var."""
+    return os.getenv("MINDEE_V2_SE_TESTS_FINDOC_MODEL_ID", "")
