+++ tests/v2/test_client.py
@@ -1,8 +1,12 @@
+import concurrent.futures
 import json
 import os
+import re
+import time
 
 import httpx
 import pytest
+import respx
 
 from mindee import ExtractionParameters, ExtractionResponse, LocalResponse
 from mindee.error.mindee_error import MindeeError
@@ -268,3 +272,38 @@ def test_client_closes_httpx_connections() -> None:
         RuntimeError, match=r"Cannot send a request, as the client has been closed\."
     ):
         client.mindee_api.http_client.get("https://google.com")
+
+
+@pytest.mark.v2
+@respx.mock
+def test_httpx_multiple_calls_thread_safety() -> None:
+    client = Client(api_key="dummy_key")
+    input_path = FILE_TYPES_DIR / "pdf" / "blank_1.pdf"
+
+    def delayed_response(request: httpx.Request) -> httpx.Response:
+        job_json = json.loads((V2_DATA_DIR / "job" / "ok_processing.json").read_text())
+        time.sleep(0.1)
+        return httpx.Response(201, json=job_json)
+
+    url_pattern = re.compile(r"https://api-v2\.mindee\.net/v2/.+/enqueue")
+    respx.post(url_pattern).mock(side_effect=delayed_response)
+
+    def make_request():
+        input_source = PathInput(input_path)
+        params = ExtractionParameters(model_id="dummy-model-id")
+        return client.enqueue(input_source, params)
+
+    thread_count = 20
+    successful_responses = 0
+    with concurrent.futures.ThreadPoolExecutor(max_workers=thread_count) as executor:
+        futures = [executor.submit(make_request) for _ in range(thread_count)]
+
+        for future in concurrent.futures.as_completed(futures):
+            response = future.result()
+            if (
+                response.job
+                and response.job.id == "12345678-1234-1234-1234-123456789ABC"
+            ):
+                successful_responses += 1
+
+    assert successful_responses == thread_count
